Rename the init commands
